@@ -1,61 +1,71 @@
 {
-    "name": "es-restful",
-    "version": "2.3.0",
-    "description": "easy & simple nodejs restful server framework",
-    "main": "lib/index.js",
-    "types": "lib/index.d.ts",
-    "files": [
-        "lib"
-    ],
-    "scripts": {
-        "start": "nodemon --watch 'example/**/*.ts' --exec 'ts-node' example/server.ts",
-        "build": "rimraf lib/* && tsc -p ./src",
-        "dev": "rimraf lib/* && tsc -p ./src -w",
-        "prepublish": "npm run build",
-        "test": "./node_modules/.bin/mocha -r ./node_modules/ts-node/register ./tests/**/**/*.ts",
-        "cover": "rimraf coverage && ./node_modules/.bin/istanbul cover ./node_modules/mocha/bin/_mocha -- ./tests/**/**/*.ts --report lcovonly --compilers ts:ts-node/register",
-        "report": "rimraf mochawesome-report && ./node_modules/.bin/mocha -r ./node_modules/ts-node/register ./tests/**/**/*.ts --recursive --reporter mochawesome",
-        "changelog": "conventional-changelog -p angular -i CHANGELOG.md -w -r 0 >> CHANGELOG.md"
+  "name": "es-restful",
+  "version": "2.3.0",
+  "description": "easy & simple nodejs restful server framework",
+  "main": "lib/index.js",
+  "types": "lib/index.d.ts",
+  "files": [
+    "lib"
+  ],
+  "scripts": {
+    "start": "nodemon --watch 'example/**/*.ts' --exec 'ts-node' example/server.ts",
+    "build": "rimraf lib/* && tsc -p tsconfig.build.json",
+    "dev": "rimraf lib/* && tsc -w",
+    "prepublish": "npm run build",
+    "test": "jest",
+    "cover": "jest --coverage",
+    "changelog": "conventional-changelog -p angular -i CHANGELOG.md -w -r 0 >> CHANGELOG.md"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/lleohao/restful.git"
+  },
+  "keywords": [
+    "nodejs",
+    "typescript",
+    "restful",
+    "web",
+    "framework",
+    "api"
+  ],
+  "author": "lleohao",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/lleohao/restful/issues"
+  },
+  "homepage": "https://github.com/lleohao/restful#readme",
+  "devDependencies": {
+    "@types/bluebird": "^3.5.24",
+    "@types/jest": "^23.3.1",
+    "@types/node": "^6.0.117",
+    "coveralls": "^2.13.3",
+    "cz-conventional-changelog": "^1.2.0",
+    "docsify": "^1.10.5",
+    "jest": "^23.5.0",
+    "nodemon": "^1.11.0",
+    "rimraf": "^2.6.2",
+    "ts-jest": "^23.1.4",
+    "ts-node": "^1.7.3",
+    "typescript": "^2.9.2"
+  },
+  "config": {
+    "commitizen": {
+      "path": "./node_modules/cz-conventional-changelog"
+    }
+  },
+  "dependencies": {},
+  "jest": {
+    "transform": {
+      "^.+\\.tsx?$": "ts-jest"
     },
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/lleohao/restful.git"
-    },
-    "keywords": [
-        "nodejs",
-        "typescript",
-        "restful",
-        "web",
-        "framework",
-        "api"
-    ],
-    "author": "lleohao",
-    "license": "MIT",
-    "bugs": {
-        "url": "https://github.com/lleohao/restful/issues"
-    },
-    "homepage": "https://github.com/lleohao/restful#readme",
-    "devDependencies": {
-        "@types/bluebird": "^3.0.37",
-        "@types/mocha": "^2.2.33",
-        "@types/node": "^6.0.51",
-        "@types/should": "^8.1.30",
-        "coveralls": "^2.11.15",
-        "cz-conventional-changelog": "^1.2.0",
-        "docsify": "^1.3.5",
-        "istanbul": "^0.4.5",
-        "mocha": "^3.2.0",
-        "mochawesome": "^1.5.4",
-        "nodemon": "^1.11.0",
-        "rimraf": "^2.6.1",
-        "should": "^11.1.1",
-        "ts-node": "^3.3.0",
-        "typescript": "^2.5.2"
-    },
-    "config": {
-        "commitizen": {
-            "path": "./node_modules/cz-conventional-changelog"
-        }
-    },
-    "dependencies": {}
+    "testRegex": "(/__tests__/.*|(\\.|/)(test|spec))\\.(jsx?|tsx?)$",
+    "moduleFileExtensions": [
+      "ts",
+      "tsx",
+      "js",
+      "jsx",
+      "json",
+      "node"
+    ]
+  }
 }
