@@ -1,208 +1,201 @@
-import { createServer, request, Server, get } from 'http';
-import { requestParse } from '../lib/requestParse';
-import { stringify } from 'querystring';
-import should = require('should');
-
-should['config'].checkProtoEql = false;
-
-const urlencoded = 1;
-const json = 2;
+import { createServer, request } from 'http';
+import { requestParse } from '../src/requestParse';
 
 describe('RequestParse', () => {
-    const haveBodyCases = {
-        jsonParserTest: {
-            herder: ['application/json'],
-            _cases: [
-                {
-                    path: '/',
-                    data: {
-                        name: 'lleohao'
-                    },
-                    e: {
-                        data: {
-                            name: 'lleohao'
-                        },
-                        rawData: null
-                    }
-                },
-                {
-                    path: '/?age=22',
-                    data: {
-                        name: 'lleohao'
-                    },
-                    e: {
-                        rawData: null,
-                        data: {
-                            age: '22',
-                            name: 'lleohao'
-                        }
-                    }
-                }
-            ]
+  const haveBodyCases = {
+    jsonParserTest: {
+      herder: ['application/json'],
+      _cases: [
+        {
+          path: '/',
+          data: {
+            name: 'lleohao'
+          },
+          e: {
+            data: {
+              name: 'lleohao'
+            },
+            rawData: null
+          }
         },
-        textParserTest: {
-            herder: ['text/palin', 'text/xml'],
-            _cases: [
-                {
-                    path: '/',
-                    data: '<author name="lleohao">',
-                    e: {
-                        data: {},
-                        rawData: '<author name="lleohao">'
-                    }
-                },
-                {
-                    path: '/?age=22',
-                    data: '<author name="lleohao">',
-                    e: {
-                        data: {
-                            age: '22'
-                        },
-                        rawData: '<author name="lleohao">'
-                    }
-                }
-            ]
-        },
-        urlencodedParserTest: {
-            herder: ['application/x-www-form-urlencoded'],
-            _cases: [
-                {
-                    path: '/',
-                    data: 'name=lleohao',
-                    e: {
-                        rawData: null,
-                        data: {
-                            name: 'lleohao'
-                        }
-                    }
-                },
-                {
-                    path: '/?age=22',
-                    data: 'name=lleohao',
-                    e: {
-                        rawData: null,
-                        data: {
-                            age: '22',
-                            name: 'lleohao'
-                        }
-                    }
-                }
-            ]
-        },
-        rawParserTest: {
-            herder: ['any/any', 'multipart/form-data'],
-            _cases: [
-                {
-                    path: '/',
-                    data: 'datatatatatattata',
-                    e: {
-                        data: {},
-                        rawData: 'datatatatatattata'
-                    }
-                },
-                {
-                    path: '/?age=22',
-                    data: 'datatatatatattata',
-                    e: {
-                        data: {
-                            age: '22'
-                        },
-                        rawData: 'datatatatatattata'
-                    }
-                }
-            ]
+        {
+          path: '/?age=22',
+          data: {
+            name: 'lleohao'
+          },
+          e: {
+            rawData: null,
+            data: {
+              age: '22',
+              name: 'lleohao'
+            }
+          }
         }
-    };
-
-    const noBodyCases = [
+      ]
+    },
+    textParserTest: {
+      herder: ['text/palin', 'text/xml'],
+      _cases: [
         {
-            method: 'GET',
-            path: '/',
-            e: {
-                data: {},
-                rawData: null
-            }
+          path: '/',
+          data: '<author name="lleohao">',
+          e: {
+            data: {},
+            rawData: '<author name="lleohao">'
+          }
         },
         {
-            method: 'GET',
-            path: '/?name=lleohao',
-            e: {
-                data: {
-                    name: 'lleohao'
-                },
-                rawData: null
-            }
-        },
-        {
-            method: 'DELETE',
-            path: '/',
-            e: {
-                rawData: null,
-                data: {}
-            }
-        },
-        {
-            method: 'DELETE',
-            path: '/?name=lleohao',
-            e: {
-                data: {
-                    name: 'lleohao'
-                },
-                rawData: null
-            }
+          path: '/?age=22',
+          data: '<author name="lleohao">',
+          e: {
+            data: {
+              age: '22'
+            },
+            rawData: '<author name="lleohao">'
+          }
         }
-    ]
+      ]
+    },
+    urlencodedParserTest: {
+      herder: ['application/x-www-form-urlencoded'],
+      _cases: [
+        {
+          path: '/',
+          data: 'name=lleohao',
+          e: {
+            rawData: null,
+            data: {
+              name: 'lleohao'
+            }
+          }
+        },
+        {
+          path: '/?age=22',
+          data: 'name=lleohao',
+          e: {
+            rawData: null,
+            data: {
+              age: '22',
+              name: 'lleohao'
+            }
+          }
+        }
+      ]
+    },
+    rawParserTest: {
+      herder: ['any/any', 'multipart/form-data'],
+      _cases: [
+        {
+          path: '/',
+          data: 'datatatatatattata',
+          e: {
+            data: {},
+            rawData: 'datatatatatattata'
+          }
+        },
+        {
+          path: '/?age=22',
+          data: 'datatatatatattata',
+          e: {
+            data: {
+              age: '22'
+            },
+            rawData: 'datatatatatattata'
+          }
+        }
+      ]
+    }
+  };
 
-    noBodyCases.forEach(({ method, path, e }, i) => {
-        it(` method: ${method}, path: ${path}`, (done) => {
-            const port = 7070 + i;
+  const noBodyCases = [
+    {
+      method: 'GET',
+      path: '/',
+      e: {
+        data: {},
+        rawData: null
+      }
+    },
+    {
+      method: 'GET',
+      path: '/?name=lleohao',
+      e: {
+        data: {
+          name: 'lleohao'
+        },
+        rawData: null
+      }
+    },
+    {
+      method: 'DELETE',
+      path: '/',
+      e: {
+        rawData: null,
+        data: {}
+      }
+    },
+    {
+      method: 'DELETE',
+      path: '/?name=lleohao',
+      e: {
+        data: {
+          name: 'lleohao'
+        },
+        rawData: null
+      }
+    }
+  ];
+
+  noBodyCases.forEach(({ method, path, e }, i) => {
+    it(` method: ${method}, path: ${path}`, done => {
+      const port = 7070 + i;
+      let server = createServer(async (req, res) => {
+        const data = await requestParse(req);
+        res.end();
+
+        expect(data).toEqual(e);
+        done();
+        server.close();
+      });
+      server.listen(port);
+
+      request({ host: '127.0.0.1', port, path, method }).end();
+    });
+  });
+
+  for (const testName in haveBodyCases) {
+    const { herder, _cases } = haveBodyCases[testName];
+
+    describe(testName, () => {
+      _cases.forEach(({ path, e, data }) => {
+        herder.forEach((h, i) => {
+          const port = 6060 + i;
+          it(`path: ${path}, header: ${h}`, done => {
             let server = createServer(async (req, res) => {
-                const data = await requestParse(req);
-                res.end();
+              const data = await requestParse(req);
+              expect(data).toEqual(e);
 
-                should(data).be.deepEqual(e);
-                done();
-                server.close();
+              done();
+              res.end();
+              server.close();
             });
             server.listen(port);
 
-            request({ host: '127.0.0.1', port, path, method }).end()
+            let req = request({ host: '127.0.0.1', port, path: path, method: 'POST' });
+            req.setHeader('Content-Type', h);
+
+            let reqData = data;
+
+            if (/json/i.test(testName)) {
+              reqData = JSON.stringify(data);
+            }
+
+            req.write(reqData);
+
+            req.end();
+          });
         });
+      });
     });
-
-    for (const testName in haveBodyCases) {
-        const { herder, _cases } = haveBodyCases[testName];
-
-        describe(testName, () => {
-            _cases.forEach(({ path, e, data }) => {
-                herder.forEach((h, i) => {
-                    const port = 6060 + i;
-                    it(`path: ${path}, header: ${h}`, (done) => {
-                        let server = createServer(async (req, res) => {
-                            const data = await requestParse(req);
-                            should(data).be.eql(e);
-
-                            done();
-                            res.end();
-                            server.close();
-                        });
-                        server.listen(port);
-
-                        let req = request({ host: '127.0.0.1', port, path: path, method: 'POST' });
-                        req.setHeader('Content-Type', h);
-
-                        let reqData = data;
-
-                        if (/json/i.test(testName)) {
-                            reqData = JSON.stringify(data);
-                        }
-
-                        req.write(reqData);
-
-                        req.end();
-                    });
-                });
-            });
-        });
-    }
+  }
 });
