@@ -2,8 +2,8 @@ import jsonParser from './jsonParser';
 import queryParser from './queryParser';
 import rawParser from './rawParser';
 import textParser from './textParser';
-import urlencodeParser from './urlencodeParser';
+import urlencodedParser from './urlencodedParser';
 
 export type parserInterface = (rawData: string, callback: (err: Error, data: {}) => void) => void;
 
-export { queryParser, urlencodeParser, jsonParser, textParser, rawParser };
+export { queryParser, urlencodedParser, jsonParser, textParser, rawParser };
