@@ -1,7 +1,7 @@
 import { parserInterface } from './index';
 
 const rawParser: parserInterface = (body, callback) => {
-    callback(null, body);
+  callback(null, body);
 };
 
 export default rawParser;
