@@ -1,7 +1,7 @@
 import { parserInterface } from './index';
 
 const textParser: parserInterface = (body, callback) => {
-    callback(null, decodeURIComponent(body));
+  callback(null, decodeURIComponent(body));
 };
 
 export default textParser;
